% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_values_to_key.R
\name{add_values_to_key}
\alias{add_values_to_key}
\title{Add Values to Key}
\usage{
add_values_to_key(values_to_add, prime_key)
}
\arguments{
\item{values_to_add}{A vector of values to add to key.}

\item{prime_key}{The old keys to which values are to be added.}
}
\value{
A vector containing a new key, including both old and new values.
}
\description{
Adds values to an encoding key as generated by \code{make_encoding_key()}.
}
\examples{
# Add "Arthropod" to encoding key.
new_prime_key <- add_values_to_key("Arthropod", c(Fish = 2, Mammal = 3, Bird = 5))
}
\seealso{
\code{\link[=make_encoding_key]{make_encoding_key()}}
}
