% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amplicons}
\alias{amplicons}
\title{Amplicons Panel}
\format{
A data.frame
\describe{
  \item{Col1}{chromosome}
  \item{Col2}{amplicon 1 start position}
  \item{Col3}{amplicon 1 end position}
  \item{Col4}{amplicon 2 start position}
  \item{Col5}{amplicon 2 end position}
  \item{Col6}{amplicon id}
  \item{Col7}{gene name}
  \item{Col8}{gene strand}
}
}
\usage{
amplicons
}
\description{
An example of an amplicon panel .txt file. Each row is an amplicon.
Columns order is: chromosome, amplicon 1 start position, amplicon 1 end position, amplicon 2 start position, amplicon 2 end position, amplicon id, gene_name, gene strand.
Order of amplicon is based on 5'-3' direction. File has no header.
}
\concept{data}
\keyword{datasets}
