% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_values.R
\name{append_values}
\alias{append_values}
\title{Append Values}
\usage{
append_values(encoded_value, values_to_append, prime_key)
}
\arguments{
\item{encoded_value}{The old encoded value, as generated by \code{encode_value()}.}

\item{values_to_append}{A vector of values to add to the encoded value.}

\item{prime_key}{An encoding key, as generated by \code{make_encoding_key()}.}
}
\value{
A double containing an integer with the new encoded value.
}
\description{
Appends values to a value that has been previously encoded by
\code{encode_value()}, using a key generated by \code{make_encoding_key()}.
}
\examples{
# Add "Arthropod" to the previously encoded value.
new_encoded_value <- append_values(6, "Arthropod", c(Fish = 2, Mammal = 3, Bird = 5, Arthropod = 7))
}
\seealso{
\code{\link[=encode_value]{encode_value()}}

\code{\link[=make_encoding_key]{make_encoding_key()}}
}
