% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{construct_objfun}
\alias{construct_objfun}
\title{Construct an objective function from a user given one.}
\usage{
construct_objfun(fun)
}
\arguments{
\item{fun}{The function given to \code{obj.fun} in the 
main function.}
}
\description{
Construct an objective function from a user given one.
}
\details{
The purpose is to create an objective function
   that takes two arguments: \code{y} for the response
   vector and \code{inbox} a logical vector indicating
   if each observation lies in the box for which the
   objective function is computed.
}
