% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countReads.R
\name{countReads}
\alias{countReads}
\title{Counts number of reads for each gene in each barcodes}
\usage{
countReads(
  out = "/path.to.ouput.dir/",
  qnames.path = "/path.to.qnames/",
  format = "both",
  sample.id = "test_small"
)
}
\arguments{
\item{out}{Output directory}

\item{qnames.path}{Path to the qnames file made with runPRIMR() function}

\item{format}{Return the counts matrix in a gene by cells format ('seurat'), in a cells by genes format ('anndata'), or both ('both')}

\item{sample.id}{Character vector of length one indicating the sample name}
}
\value{
The filtered counts matrix
}
\description{
Counts number of reads for each gene in each barcodes
}
