% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_value.R
\name{decode_value}
\alias{decode_value}
\title{Decode Value}
\usage{
decode_value(values_to_decode, prime_key)
}
\arguments{
\item{values_to_decode}{A product of prime numbers multiplied using
\code{encode_value()}.}

\item{prime_key}{A key as generated by \code{make_encoding_key()}}
}
\value{
A vector of decoded values.
}
\description{
Takes a product from \code{encode_value()} and returns a vector of decoded values
using a key as generated by \code{make_encoding_key()}.
}
\examples{
# Return a vector containing "Mammal" and "Bird".
decoded_values <- decode_value(15, c(Fish = 2, Mammal = 3, Bird = 5))
}
\seealso{
\code{\link[=encode_value]{encode_value()}}

\code{\link[=make_encoding_key]{make_encoding_key()}}
}
