% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_value.R
\name{encode_value}
\alias{encode_value}
\title{Encode Value}
\usage{
encode_value(values_to_encode, prime_key)
}
\arguments{
\item{values_to_encode}{A vector of values to encode.}

\item{prime_key}{An encoding key, as generated by \code{make_encoding_key()}.}
}
\value{
A double, containing an integer value.
}
\description{
Takes a vector and returns an integer that is a product of the prime keys
corresponding to the values in the vector. The second argument takes an
encoding key as generated by \code{make_encoding_key()}. Values can be decoded
using \code{decode_value()}.
}
\examples{
# Generate a product equal to 15.
encoded_value <- encode_value(c('Mammal', 'Bird'), c(Fish = 2, Mammal = 3, Bird = 5))
}
\seealso{
\code{\link[=make_encoding_key]{make_encoding_key()}}

\code{\link[=decode_value]{decode_value()}}
}
