% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_encoding_key.R
\name{make_encoding_key}
\alias{make_encoding_key}
\title{Make Encoding Key}
\usage{
make_encoding_key(values_to_encode)
}
\arguments{
\item{values_to_encode}{A vector containing all unique values to encode.}
}
\value{
A vector of prime numbers with names corresponding to the values to
be encoded.
}
\description{
Takes a vector of values and pairs them with prime numbers. The pairs can
then be used for encoding by \code{encode_value()}, decoding by \code{decode_value()}.
If there are several identical values in the input vector, only one prime
number per unique value is returned.
}
\examples{
# Return a key of three prime numbers for three categories.
prime_key <- make_encoding_key(c("Mammal", "Bird", "Fish"))
}
\seealso{
\code{\link[=encode_value]{encode_value()}}

\code{\link[=decode_value]{decode_value()}}
}
