% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Plot a peeling trajectory}
\usage{
plot_trajectory(object, xtype = c("support", "nobs"), ytype = c("yfun",
  "diff", "rel.diff"), npeel = NULL, support = NULL, yfun = NULL,
  npaste = NULL, abline.pars = list(), se = TRUE, se.pars = list(),
  ...)
}
\arguments{
\item{object}{A \code{prim} object.}

\item{xtype}{A character indicating the how to display the x axis of
the plot. 'support' (the default) for the support of the boxes or
'nobs' for the number of observations inside the box.}

\item{ytype}{A character indicating which trajectory to plot. 'yfun'
(the default) for the objective function value, 'diff' for the
difference between successive boxes and 'rel.diff' for the relative
difference. See \code{\link{jump.prim}} for details.}

\item{npeel}{Integer vector. If not \code{NULL}, draw a vertical line 
at the corresponding peeling iterations of the trajectory.}

\item{support}{Numeric vector. If not \code{NULL}, draw a vertical line 
at the corresponding supports of the trajectory.}

\item{yfun}{Numeric vector. If not \code{NULL}, draw a vertical line 
at the boxes with the closest \code{yfun} value.}

\item{npaste}{Integer vector. If not \code{NULL}, draw a vertical line 
at the corresponding pasting iterations of the trajectory.}

\item{abline.pars}{List of parameters to be passed to
\code{\link[graphics]{abline}} for the vertical lines.}

\item{se}{Logical indicating if standard error bars should be drawn.
Works only for \code{cv.prim} objects and \code{ytype = "yfun"}.}

\item{se.pars}{Parameters to be passed to \code{arrows} for drawing
strandard error bars.}

\item{...}{Additional graphical parameters for 
\code{\link[graphics]{plot}}.}
}
\description{
Displays the peeling trajectory of a \code{prim} object with chosen
   cut-points.
}
\examples{
   # A simple bump
   set.seed(12345)
   x <- matrix(runif(2000), ncol = 2, dimnames = list(NULL, c("x1", "x2")))
   y <- 2 * x[,1] + 5 * x[,2] + 10 * (x[,1] >= .8 & x[,2] >= .5) + 
     rnorm(1000)
   # Peeling with alpha = 0.05 and beta.stop = 0.05
   peel_res <- peeling(y, x, beta.stop = 0.05)

   # Display the peeling trajectory
   plot_trajectory(peel_res, type = "b", pch = 16, col = "cornflowerblue", 
     support = 0.11, abline.pars = list(lwd = 2, col = "indianred"), 
     xlab = "")

}
\references{
Friedman, J.H., Fisher, N.I., 1999. Bump hunting in high-dimensional data.
     Statistics and Computing 9, 123-143. 
     https://doi.org/10.1023/A:1008894516817
}
\seealso{
\code{\link{peeling}} for peeling trajectories.
}
