% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_values.R
\name{remove_values}
\alias{remove_values}
\title{Remove Values}
\usage{
remove_values(encoded_value, values_to_remove, prime_key)
}
\arguments{
\item{encoded_value}{An encoded value, as encoded by \code{encode_value()}.}

\item{values_to_remove}{A vector containing values to remove from encoded value.}

\item{prime_key}{An encoding key, as generated by \code{make_encoding_key()}.}
}
\value{
A double, containing an integer, where the values to remove has been removed.
}
\description{
Removes values from an encoded value.
}
\examples{
# Remove "Fish" from encoded value.
new_encoded_value <- remove_values(6, "Fish", c(Fish = 2, Mammal = 3, Bird = 5, Arthropod = 7))
}
\seealso{
\code{\link[=encode_value]{encode_value()}}

\code{\link[=make_encoding_key]{make_encoding_key()}}
}
