% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPRIMR.R
\name{runPRIMR}
\alias{runPRIMR}
\title{Split the original BAM file into amplicons using a amplicon file}
\usage{
runPRIMR(
  out = "/gpfs/commons/home/tbotella/PRIMR/tests",
  bamfile = "/gpfs/commons/home/tbotella/PRIMR/data/datatest.bam",
  ampfile = "/gpfs/commons/home/tbotella/PRIMR/data/ampliconsPanel.txt",
  sample.id = "testSample",
  nthreads = 4
)
}
\arguments{
\item{out}{Output directory}

\item{bamfile}{Path to the aligned BAM file}

\item{ampfile}{Path to the amplicon file. The file is a .txt file (tabulation) without header. The columns order is : chr, amplicon1_start, amplicon1_end, amplicon2_start, amplicon2_end, amplicon_id, gene_name and gene_strand (+ if gene's coding strand is on the forward strand, - if it's on the reverse strand 3'-5')}

\item{sample.id}{Character vector of length one indicating the sample name}

\item{nthreads}{Integer indicating the number of threads to use for parallel processing}
}
\value{
Create a folder with the reads that pass PRIMR QC, a filtered BAM file, a barplot as pdf for read counts pre/post filtering and a tmp folder with all intermediate files
}
\description{
Split the original BAM file into amplicons using a amplicon file
}
